requirement.txt
app==0.0.1
Appium-Python-Client==0.26
asyncio==3.4.3
certifi==2018.1.18
chardet==3.0.4
idna==2.6
PyYAML==3.12
requests==2.18.4
selenium==3.11.0
urllib3==1.22
